refactor: un poco de limpieza de codigo
