+++ src/ui/components/game-object.model.ts
